Developed DataBase H2 with Jpa

+++ src/main/java/com/rosineimartins/projetoprodutos/controller/ProdutoController.java
@@ -1,7 +1,6 @@
 package com.rosineimartins.projetoprodutos.controller;
 
 import com.rosineimartins.projetoprodutos.model.Produto;
-import com.rosineimartins.projetoprodutos.model.exception.ResourceNotFoundException;
 import com.rosineimartins.projetoprodutos.service.ProdutoService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -26,18 +25,18 @@ public class ProdutoController {
     }
 
     @GetMapping("/{id}")
-    public Optional<Produto> obterPorId(@PathVariable Integer id) {
+    public Optional<Produto> obterPorId(@PathVariable Long id) {
         return produtoService.obterPorId(id);
     }
 
     @DeleteMapping("/{id}")
-    public String deletarProduto(@PathVariable Integer id) {
+    public String deletarProduto(@PathVariable Long id) {
         produtoService.deletarProduto(id);
         return "Produto com id: " + id + " deletado com sucesso";
     }
 
     @PutMapping("/{id}")
-    public Produto atualizar(@RequestBody Produto produto, @PathVariable Integer id) {
+    public Produto atualizar(@RequestBody Produto produto, @PathVariable Long id) {
         return produtoService.atualizar(id,produto);
     }
 }
